add 'ship edit' command
read symlinked files for gogetter local upstreams

persist replicated.app to state

refactor 'replicatedapp' and 'state' packages to resolve import cycle

remove entitlements info and duplicate license info from persisted app release

entitlements info should be present
