+++ pkg/cli/root.go
@@ -81,6 +81,7 @@ func RootCmd() *cobra.Command {
 	cmd.AddCommand(Init())
 	cmd.AddCommand(Watch())
 	cmd.AddCommand(Update())
+	cmd.AddCommand(Edit())
 	cmd.AddCommand(Unfork())
 	cmd.AddCommand(App())
 	cmd.AddCommand(Version())
